remove redundant extern primes; incorrect print format

+++ test/unit/test_mul_csa_pthreads.c
@@ -59,7 +59,7 @@ static void print_limbs(const char* tag, const cfx_big_t* x, size_t upto) {
     fprintf(stderr, "%s: [", tag);
     for (size_t i = 0; i < upto; ++i) {
         cfx_limb_t v = (i < x->n) ? x->limb[i] : 0;
-        fprintf(stderr, "%s" CFX_PRIuLIMB "", (i ? ", " : ""), (unsigned long long)v);
+        fprintf(stderr, "%s" CFX_PRIuLIMB "", (i ? ", " : ""), v);
     }
     fprintf(stderr, "]\n");
 }
